detection/requirements.txt
numpy
scikit-learn
psycopg2-binary

